requirements.txt
flask>=2.0.0
pillow>=9.5.0
numpy>=1.21.0
requests>=2.25.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
trimesh>=3.20.0
scipy>=1.10.0
gunicorn>=20.1.0 